tweaked scroll-to-top behavior for paged fragments
As of this change, paged fragments have a floating action button to scroll to top. Instead of relying on loadItemsInitially for paged fragments, this value is now determined by whether or not the current fragment is active in the relevant FragmentStatePagerAdapter.

If it is not, the FAB is not changed. If the current fragment is active, we set the click listener to scroll up in the current fragment. Beforehand, a FragmentStatePagerAdapter could have changed the scroll click listener even if it wasn't the active fragment, often appearing as "does nothing".