removed most of the trading section except for feedback on user profiles
As per https://www.steamgifts.com/discussion/yWGlf/steamtrades SteamTrades
is essentially handled as separate site. This is noticable by two different things:

- usernames on SteamTrades are not necessarily identical to SteamGifts ones.
  This is mitigated by using the SteamID64 for trade feedback user links.
  Keep in mind trade feedback isn't necessarily feature-complete.
- A lot of HTML class names changed.

Meanwhile, there is no sight of an official API coming ever. And, quite
frankly, with random undocumented website changes (see a1adfe24), this is a
huge pain in the ass.

Contributors are of course welcome to (partially) revert this change to restore
the trade list in a pull request. Issues that are blatantly obvious:

- You need to login separately on SteamTrades, which may not even need to be
  the same user (you *might* violate the SteamTrades Terms of Services in that
  case, but I haven't read them)
- Deleted comments are not handled.
- HTML/CSS class names are entirely different for the trades, comments, etc.
  from SteamGifts.
- User profiles on SteamGifts may not exist for a SteamTrades user.
- Replies to negative reviews are not checked and may only be displayed by
  pure chance. I've never seen a reply.
- "Safety" information such as VAC or game bans, user registration, last login,
  steam level, game count and SteamID/*32/*64 are not displayed.
- Your own steam trades are shown by visiting the site
  https://www.steamtrades.com/trades/search?user=yoursteamid64
  but you can't directly fetch the SteamID64 from SteamGifts, thus proving
  an annoyance in building the navigation.

Your milage may vary, but for a feature I've never used I can't justify the
effort for myself. Packing both "fixing SteamTrades comments on profiles"
and "deleting the rest of SteamTrades" may be an unfortunately big comment,
but that's how it is.